+++ backend/tests/test_intraday_ticker.py
@@ -1,1 +1,1 @@
-from real_time_trading.intraday_ticker import IntradayTicker
+from real_time_trading.objects.intraday_ticker import IntradayTicker
